✨ Refactor fetch and mutation hooks to improve type definitions and structure
