+++ packages/client/src/types/hooks/fetch.ts
@@ -1,0 +1,7 @@
+export interface UseFetchResult<TData> {
+  data: TData | null;
+  error: Error | null;
+  isPending: boolean;
+  isSuccess: boolean;
+  isError: boolean;
+}
